requirements.txt
streamlit==1.33.0
nltk==3.8.1
python-dotenv==1.0.0
python-dotenv==1.0.0
spotipy==2.23.0